refactor(translators-interpreters): simplify code by removing unnecessary code
This commit removes unnecessary code in the `translators-interpreters` module. Specifically, the `translatorsAndInterpreters` property in the `SwornOutputTypes` interface was removed since it was not being used. Additionally, the `length` property was removed from the `props.interpreterServices` and `props.translationSpecialties` arrays in the `findPublishedTranslatorsInterpretersPerCountry` function since they are already arrays and do not need to be checked for length. Finally, the `hasSworn.translatorsAndInterpreters` condition was removed from the `translators-interpreters-results-list.njk` file since it was redundant.