refactor(translators-interpreters-results-list.njk): simplify and improve warning message
The warning message in the translators-interpreters-results-list.njk file has been simplified and improved to make it more concise and easier to understand. The message now uses a single set of conditional statements to determine whether the translators and/or interpreters listed in the results are authorized to carry out legal or official services. The message also includes additional information about the types of authorized translators and/or interpreters that may be listed in the results.