chore(lists): send email to admins when provider details are edited
Add a new function `sendProviderChangeDetailsEmailToAdmins` to send an email to
all the admins of a list when a provider's details are edited. The email is sent
using the `sendProviderChangedDetailsEmail` function from `govuk-notify` service.

Also, add a new template ID for the provider details edit email in the `NOTIFY`
configuration object.