chore(migration.sql): add migration to move list item audits to events table
This commit adds a migration script to move list item audits to the events table. The script inserts rows into the "Event" table based on the data from the "Audit" table. It selects the "createdAt" timestamp, "itemId" from the "jsonData" field, "auditEvent" type, and constructs the "jsonData" field for the "Event" table. It also includes additional data for the "reference", "notes", and "eventName" fields. After the data is moved, the script deletes all list item audit events from the "Audit" table.